Various fixes and PR updates
