+++ proknow-sdk-test/PatientTest/PatientScorecardsTest.cs
@@ -42,7 +42,6 @@ namespace ProKnow.Patient.Test
 
             // Create a test patient with a dose
             var patientItem = await TestHelper.CreatePatientAsync(_testClassName, testNumber, Path.Combine("Becker^Matthew", "RD.dcm"));
-            var entitySummary = patientItem.FindEntities(e => e.Type == "dose")[0];
 
             // Create entity scorecards object
             var entityScorecards = new PatientScorecards(_proKnow, workspace.Id, patientItem.Id);
@@ -118,7 +117,6 @@ namespace ProKnow.Patient.Test
 
             // Create a test patient with a dose
             var patientItem = await TestHelper.CreatePatientAsync(_testClassName, testNumber, Path.Combine("Becker^Matthew", "RD.dcm"));
-            var entitySummary = patientItem.FindEntities(e => e.Type == "dose")[0];
 
             // Create entity scorecards object
             var entityScorecards = new PatientScorecards(_proKnow, workspace.Id, patientItem.Id);
@@ -169,7 +167,6 @@ namespace ProKnow.Patient.Test
 
             // Create a test patient with a dose
             var patientItem = await TestHelper.CreatePatientAsync(_testClassName, testNumber, Path.Combine("Becker^Matthew", "RD.dcm"));
-            var entitySummary = patientItem.FindEntities(e => e.Type == "dose")[0];
 
             // Create entity scorecards object
             var entityScorecards = new PatientScorecards(_proKnow, workspace.Id, patientItem.Id);
@@ -217,7 +214,6 @@ namespace ProKnow.Patient.Test
 
             // Create a test patient with a dose
             var patientItem = await TestHelper.CreatePatientAsync(_testClassName, testNumber, Path.Combine("Becker^Matthew", "RD.dcm"));
-            var entitySummary = patientItem.FindEntities(e => e.Type == "dose")[0];
 
             // Create entity scorecards object
             var entityScorecards = new PatientScorecards(_proKnow, workspace.Id, patientItem.Id);
@@ -294,7 +290,6 @@ namespace ProKnow.Patient.Test
 
             // Create a test patient with a dose
             var patientItem = await TestHelper.CreatePatientAsync(_testClassName, testNumber, Path.Combine("Becker^Matthew", "RD.dcm"));
-            var entitySummary = patientItem.FindEntities(e => e.Type == "dose")[0];
 
             // Create entity scorecards object
             var entityScorecards = new PatientScorecards(_proKnow, workspace.Id, patientItem.Id);
