+++ proknow-sdk-test/ProKnowApiTest.cs
@@ -84,5 +84,15 @@ namespace ProKnow.Test
             Assert.IsFalse(connectionStatus.IsValid);
             Assert.AreEqual("Invalid or missing credentials", connectionStatus.ErrorMessage);
         }
+
+        [TestMethod]
+        public async Task GetConnectionStatusAsync_BogusCredentials()
+        {
+            var credentialsFile = Path.Combine(TestSettings.TestDataRootDirectory, "./bogus_credentials.json");
+            var proKnowApi = new ProKnowApi(TestSettings.BaseUrl, credentialsFile);
+            var connectionStatus = await proKnowApi.GetConnectionStatusAsync();
+            Assert.IsFalse(connectionStatus.IsValid);
+            Assert.AreEqual("Invalid or missing credentials", connectionStatus.ErrorMessage);
+        }
     }
 }
